+++ src/components/signup/step/Plan.tsx
@@ -5,12 +5,13 @@ import styled from '@emotion/styled';
 import Divider from '@components/common/Divider';
 import Button from '@components/common/Button';
 import { useSignupStore } from 'stores/useSignupStore';
-import { usePlanStore } from 'stores/usePlanStore';
+import { IPlan, usePlanStore } from 'stores/usePlanStore';
 import { useEffect, useState } from 'react';
 import PlanTabs from '../plan/PlanTabs';
 import PlanOption from '../plan/PlanOption';
 import MemberOption from '../plan/MemberOption';
 import BuillingOption from '../plan/BuillingOption';
+import { TStatus } from '@components/common/Status';
 
 function SignupPlan() {
   const [platformId, setPlatformId] = useState(101);
@@ -23,6 +24,17 @@ function SignupPlan() {
   const getPlan = (platformId: number) => {
     return plans.filter(plan => plan.platformId === platformId)[0];
   };
+
+  const getServiceStatus = (plan: IPlan): TStatus => {
+    const { planId, isGroup, groupMembers, isYearlyPay, billingMonth, billingDay } = plan;
+    return planId &&
+      (isGroup ? groupMembers : true) &&
+      (isYearlyPay ? billingMonth && billingDay : billingDay)
+      ? 'success'
+      : 'error';
+  };
+
+  const fulfilledPlans = plans.filter(plan => getServiceStatus(plan) === 'success');
 
   const handleSubmit = () => {
     setPlatforms([...plans]);
@@ -39,7 +51,7 @@ function SignupPlan() {
       <SignupContainer>
         <ContentWrap>
           <SignupTitle>선택한 구독 서비스별 요금제를 입력해 주세요.</SignupTitle>
-          <PlanTabs setPlatformId={setPlatformId} />
+          <PlanTabs setPlatformId={setPlatformId} getServiceStatus={getServiceStatus} />
           <div>
             <PlanOption plan={getPlan(platformId)} platformId={platformId} />
             <Divider />
@@ -48,7 +60,15 @@ function SignupPlan() {
             <BuillingOption plan={getPlan(platformId)} platformId={platformId} />
           </div>
         </ContentWrap>
-        <Button text="완료" onClick={handleSubmit} />
+        <Button
+          text={
+            fulfilledPlans.length === plans.length
+              ? '완료'
+              : `요금제를 선택해주세요. (${fulfilledPlans.length}/${plans.length})`
+          }
+          disabled={fulfilledPlans.length !== plans.length}
+          onClick={handleSubmit}
+        />
       </SignupContainer>
     </>
   );
